issue450, issue98: check vpi feature presence.

+++ testsuite/gna/issue450/testsuite.sh
@@ -2,20 +2,24 @@
 
 . ../../testenv.sh
 
-if [ "$OS" = "Windows_NT" ]; then
-    vpi_lib=`$GHDL --vpi-library-dir | sed -e 's!\\\\!/!g' -e 's!^C:!/C!g'`
-    echo vpi_lib: $vpi_lib
-    PATH="$PATH:$vpi_lib"
-fi
-
-$GHDL --vpi-compile -v gcc -c vpi2.c
-$GHDL --vpi-link -v gcc -o vpi2.vpi vpi2.o
-
 analyze disptree.vhdl
-elab_simulate disptree --vpi=./vpi2.vpi | tee disptree.out
-diff --strip-trailing-cr -q disptree.ref disptree.out
+elab disptree
 
-rm -f vpi2.o vpi2.vpi disptree.out
+if ghdl_has_feature disptree vpi; then
+  if [ "$OS" = "Windows_NT" ]; then
+      vpi_lib=`$GHDL --vpi-library-dir | sed -e 's!\\\\!/!g' -e 's!^C:!/C!g'`
+      echo vpi_lib: $vpi_lib
+      PATH="$PATH:$vpi_lib"
+  fi
+
+  $GHDL --vpi-compile -v gcc -c vpi2.c
+  $GHDL --vpi-link -v gcc -o vpi2.vpi vpi2.o
+
+  simulate disptree --vpi=./vpi2.vpi | tee disptree.out
+  diff --strip-trailing-cr -q disptree.ref disptree.out
+
+  rm -f vpi2.o vpi2.vpi disptree.out
+fi
 
 clean
